Allow reusing the same allocation in the reader loop
