+++ include/ublox.h
@@ -7,6 +7,7 @@
  * This code assumes a Little Endian host.
  */
 
+#include <o2s/array.h>        // array_t
 #include <o2s/input_stream.h> // istream_t
 #include <o2s/serial.h>       // serial_port_t
 #include <o2s/string.h>       // string_t
@@ -49,12 +50,24 @@ struct ublox_footer
 };
 
 bool             ublox_port_config(serial_port_t* port, int64_t baudrate);
-
-ublox_message_t* ublox_next_message(istream_t* port);
-
 ublox_checksum_t ublox_compute_checksum(const ublox_message_t* message, size_t size);
-
 string_t         ublox_header_tostring(const struct ublox_header* message);
+
+enum parser_error
+{
+    PARSER_SUCCESS, /**< Not an error: an ublox message was successfully parsed */
+    /* Warnings: */
+    PARSER_GARBAGE, /**< The first 2 characters read were not as expected, just try again */
+    /* Errors: */
+    PARSER_ERROR_INVALID_CLASS = 4, /**< A possible ublox message had an unknown class */
+    PARSER_ERROR_WRONG_CHECKSUM,    /**< The integrity check failed, the message may be corrupted */
+    /* Fatal errors: */
+    PARSER_ERROR_READ = 8,    /**< Could not read enough characters */
+    PARSER_ERROR_UNREACHABLE, /**< It should not be possible to emit this error */
+};
+
+enum parser_error ublox_parse_single_message(istream_t* input, array_t* output);
+bool              ublox_next_message(istream_t* input, array_t* output);
 
 /**
 @var ublox_header::class
